turn on debug for vagrant

+++ web3/apps/sites/helpers.py
@@ -125,8 +125,9 @@ def check_nginx_config():
 
 
 def flush_permissions():
-    with open("/proc/net/rpc/auth.unix.gid/flush", "w") as f:
-        f.write(str(int(time.time())))
+    if os.path.isfile("/proc/net/rpc/auth.unix.gid/flush"):
+        with open("/proc/net/rpc/auth.unix.gid/flush", "w") as f:
+            f.write(str(int(time.time())))
     Popen("/usr/sbin/nscd -i group".split())
     Popen("/usr/sbin/nscd -i passwd".split())
